ScoreBoard v0.31
Moving from listviews to recyclerviews.
Moved all old classes into a package called "old" for the time being. Will delete when move to new recyclerviews and appropriate adapters and fragments is finished.
Switching adapters to use recyclerviews.
Implementing an MVVM architecture with viewmodels and repositories accessing data and passing it to the views.
Begun creating the database and retrieving data from there.