ScoreBoard v0.36
Changed gametype to playmode and teams to teamoptions in BoardGame class.
Changed playmode from string to a PlayMode enum.
Changed teamoptions from string to a TeamOption enum.
Added TypeConverters so PlayMode and TeamOption enums can be entered into and taken out of the database.
Implemented a boardgame add function that adds a board game without categories.