ScoreBoard v0.52
Added in group category skill rating table to start tracking group members' skill ratings in different categories.
Added in DAO to access it.
Added in view that joins the skill rating table and members table based on who currently exists as a member in the group. This way if a player leaves a group and then rejoins later, their skill ratings are still available. However, while they are gone they are hidden.